Portability fixes

+++ colorcoding.cpp
@@ -4,6 +4,7 @@
 #include <string>
 
 #include <math.h>
+#include <stdio.h>
 #include <time.h>
 
 #ifdef __unix__
@@ -200,8 +201,9 @@ int main(int argc, char *argv[]) {
     PathSet paths = lightest_path(problem, num_paths, max_common, mode, max_lb_edges);
     double stop = timestamp();
     if (stats_only) {
-	printf("%15.2f %6zu %12.8f %12.8f\n", stop - start,
-	       peak_mem_usage / 1024 / 1024, paths.best_weight(), paths.worst_weight());
+	printf("%15.2f %6lu %12.8f %12.8f\n", stop - start,
+	       (unsigned long) (peak_mem_usage / 1024 / 1024),
+	       paths.best_weight(), paths.worst_weight());
     } else {
 	for (PathSet::it i = paths.begin(); i != paths.end(); ++i) {
 	    std::cout << i->path_weight();
